[fix] convert plot output paths to posix as well
We had to conver the input paths to rule make_report to create legal R string literals on Windows.
That, however, broke the connection between make_report inputs and plot_* outputs because Snakemake does string matching, not path matching.
A better solution would be to convert the paths in the run section and then use `shell` but I don't want to introduce this approach to the tutorial.